expense views WIP
had to fix some firebase functions so am updating to new Develop.

+++ Frugl/Frugl/Views/Itemization/CreateExpense/CreateExpenseViewModel.swift
@@ -16,26 +16,20 @@ class CreateExpenseViewModel {
     // MARK: - Properties
     private weak var delegate: CreateExpenseViewModelDelegate?
     private var service: FireBaseSyncable
-    var expenses: [Expense] = []
+    var expense: Expense?
     
     init(delegate: CreateExpenseViewModelDelegate, service: FireBaseSyncable = FirebaseService()) {
         self.delegate = delegate
         self.service = service
     }
     
-    func createExpense(with expense: Expense) {
-        service.saveExpense(expense: expense) { result in
-            switch result {
-            case .success(_):
-                self.expenses.append(expense)
-                self.delegate?.expenseCreatedSuccessfully()
-            case .failure(let failure):
-                print(failure.localizedDescription)
-            }
-        }
+    func createExpense(name: String, amount: Double) {
+        guard let expense = expense else { return }
+    
     }
     
-    func deleteExpense(with expense: Expense) {
+    func deleteExpense() {
+        guard let expense = expense else { return }
         service.deleteExpense(expense: expense) { result in
             switch result {
             case .success(_):
